Former-commit-id: 293cbf3e7b83276353b264d6470db73c1b1125b4

+++ modules/sofa/component/mapping/BeamLinearMapping.inl
@@ -306,7 +306,7 @@ const sofa::defaulttype::BaseMatrix* BeamLinearMapping<TIn, TOut>::getJ()
     if (matrixJ.get() == 0 || updateJ)
     {
         updateJ = false;
-        if (matrixJ.get() == 0 || matrixJ->rowBSize() != outStateSize || matrixJ->colBSize() != inStateSize )
+        if (matrixJ.get() == 0 || (unsigned int)matrixJ->rowBSize() != outStateSize || (unsigned int)matrixJ->colBSize() != inStateSize )
         {
             matrixJ.reset(new MatrixType(outStateSize * NOut, inStateSize * NIn));
         }
